requirements.txt
# pip install -r requirements.txt
selenium
webdriver-manager
beautifulsoup4
pytube
ffmpeg-python
argparse
